% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardscore.R
\name{completecase}
\alias{completecase}
\title{score test with missing data}
\usage{
completecase(x, y, s, family = c("gaussian", "binomial"))
}
\arguments{
\item{x}{vector of covariates}

\item{y}{outcome of interest. Quantitative for family="gaussian". For family="binomial" should be either a factor with two levels}

\item{s}{a covariate of interets, allowed to be missing}

\item{family}{response type. See above}
}
\description{
score test with missing data
}
