% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic.R
\name{modelselect}
\alias{modelselect}
\title{BIC model selection}
\usage{
modelselect(s, x, w, r, lambda = NULL)
}
\arguments{
\item{s}{a covariate of interets, allowed to be missing}

\item{x}{vector of covariates}

\item{w}{potentially high-dimensional vector of covariates}

\item{r}{indicator of whether s is observed}

\item{lambda}{A user supplied lambda sequence}
}
\description{
BIC model selection
}
