% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoretest.R
\name{scoretest}
\alias{scoretest}
\title{score test with missing data for continuous and binary outcomes}
\usage{
scoretest(x, y, s, w = NULL, r, family = c("gaussian", "binomial"))
}
\arguments{
\item{x}{vector of covariates}

\item{y}{outcome of interest. Quantitative for family="gaussian". For family="binomial" should be either a factor with two levels}

\item{s}{a covariate of interets, allowed to be missing}

\item{w}{potentially high-dimensional vector of covariates, if null, mean model}

\item{r}{indicator of whether s is observed}

\item{family}{response type. See above}
}
\description{
score test with missing data for continuous and binary outcomes
}
