% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoretest_supre.R
\name{scoretest_supre}
\alias{scoretest_supre}
\title{score test with missing data for time-to-event data}
\usage{
scoretest_supre(x, y, s, w, r, delta, Z)
}
\arguments{
\item{x}{matrix of covariates}

\item{y}{survival outcome}

\item{s}{a covariate of interets, allowed to be missing}

\item{w}{potentially high-dimensional vector of covariates}

\item{r}{indicator of whether s is observed}

\item{delta}{censoring indicator (1 = event, 0 = censored)}

\item{Z}{matrix of permutation data for hypothesis testing,
can be generated by Z = mvrnorm(500000, mu = c(0,0), Sigma = matrix(c(1,0,0,1), nrow = 2))}
}
\value{
A list containing p-values for supremum test, proportional hazards, and proportional odds models.
}
\description{
This function computes the p-values for proportional hazards and proportional odds models using a supermum score test.
}
